x,y werden nun korrekt als Double weiter gegeben

+++ src/commonMain/kotlin/engineEmi/Bodies/Circle.kt
@@ -30,7 +30,12 @@ class Circle(
     var strokeColor: RGBA = Colors.BLUE,
     var strokeThickness: Double = 0.0
 ) : Ebody(
-    x = x, y = y, density = density, friction = friction, restitution = restitution, bodyType = bodyType
+    x = x.toDouble(),
+    y = y.toDouble(),
+    density = density,
+    friction = friction,
+    restitution = restitution,
+    bodyType = bodyType
 ) {
